requirements.txt
Appium-Python-Client==3.1.1
pytest==8.0.0
selenium==4.17.2